[eventhubs] - Upgrade Event Hubs to core-tracing preview.14 (#20240)
### Packages impacted by this PR
@azure/event-hubs
@azure/test-utils

### Issues associated with this PR
#20213 

### Describe the problem that is addressed by this PR
Now that @azure/core-tracing@preview.14 is out, and hopefully the last version before GA, we need to upgrade a few packages in order to dogfood both the upgrade experience and the usage of these packages with the new [instrumentation package](https://www.npmjs.com/package/@azure/opentelemetry-instrumentation-azure-sdk).

My goal was to upgrade one AMQP package and a few HTTP packages in addition to core-rest-pipeline to collect feedback.

Upgrading EventHubs now allows us to start using the new APIs in a client package.

### What are the possible designs available to address the problem? If there are more than one possible design, why was the one in this PR chosen?
Haven't decided if this should be merged _or_ if we should just publish to NPM under a custom tag in order to allow customers to dogfood. I know there's an internal IoT Hub engineer that is interested in helping out so we just need something for them to try. 

Right now though I am leaning towards merging this in.

I also decided to dump all the `parentSpan` legacy stuff - there really isn't a need to maintain tracing backwards compat with an option that has not been used in over 6 months.