Fixing cspell errors (#20310)
Spell check scanning of packages listed above detected spelling errors in their public API surface. Many of those errors are false positives, which have been added to the `.vscode/cspell.json` file in order to ignore them. However, since all of those packages are already in GA, we cannot make changes to fix the actual spelling errors. For this reason, all of them have been listed in the file to be ignored as well.