[KeyVault] - Fix recorded browser tests (#20412)
### Packages impacted by this PR
@azure/keyvault-keys
@azure/keyvault-certificates
@azure/keyvault-secrets

Note: @azure/keyvault-admin is _not_ impacted because it has no browser test.

### Describe the problem that is addressed by this PR
I noticed that our recorded browser tests were failing in the browser. My good
friend `git bisect` pointed to 9afbe3fc46b63516e88b508911b2b2d4517dd42b as the
commit that introduced this issue.

Digging into it, it looks like this is unique to KeyVault in that it is _still_
on coreV1 but uses Identity which is on coreV2. Not sure about other client
libraries, it looks like they were fixed in the PR itself.

### What are the possible designs available to address the problem? If there are more than one possible design, why was the one in this PR chosen?
Really we just need to migrate KV to coreV2 and the new recorder... but this
will unblock the tests for now

### Provide a list of related PRs _(if any)_
https://github.com/Azure/azure-sdk-for-js/pull/20201