[ServiceBus] fix infinite retry cycles
The live test shows that with certain timing we could get into a situation where
after the connection is closed, we throw an error in
`_initAndAddCreditOperation`. and we keep restarting retry cycles in the
`forEverRetry()` loop. User calling `close()` wouldn't help because the place we
check for suspended state and throw `AbortError` is after we check the
connection and throw error.

This PR moves the checks of whether the user has suspended the receiver so we
could throw `AbortError` to end the retry cycles.

I also fix the issue of not respecting 0 values in user specified retry options.