Require test agent and phone number for live tests
When running the update capabilities live tests, if the AZURE_TEST_AGENT and an AZURE_PHONE_NUMBER_<TEST_AGENT> values are missing, tests no longer fallback to the AZURE_PHONE_NUMBER value.

This means that if at least one of those values are undefined, then the test run will fail due to not having a valid phone number to test with.