[ServiceBus] Respect user request to close receiver when disconnected (#20427)
When receiving messages in a stream using `receiver.subscribe()`, 
Under certain situation when the underlying connection is closed, we don't respect user request to close the receiver, but keep retrying to recover.

The problem is that we check the connection status and throw an error if it is closed in
`_initAndAddCreditOperation`.  We restart retry cycle because of the error in the
`retryForEver()` function with the same operation, trying to re-establish the connection, then throw the same error again.
User calling `close()` wouldn't help because the place we
check for suspended state then throw `AbortError` is after we check the
connection and throw error.

This PR adds the checking of whether user has suspended the receiver before checking of connection so we
could throw `AbortError` to end the retry cycles.  The check after `preInitialize()` invocation is still needed
as user could suspend us in the `preInitialize()` handler.

I also fix another issue of not respecting 0 values in user specified retry options.

### Packages impacted by this PR

@azure/service-bus

### Issues associated with this PR

#20388