[instrumentation] - Add integration tests (#20462)
### Packages impacted by this PR
@azure/opentelemetry-instrumentation-azure-sdk

### Issues associated with this PR
Resolves #18738

### Describe the problem that is addressed by this PR
This PR solves the following:
1. I would like to have an end-to-end test that uses our pipeline and ensures
tracing works as expected. The idea here is to ensure we do not regress in the
future when making core changes. I added a node-only integration test that uses
core-rest-pipeline and a fake client.
1. We currently maintain a TestTracer, TestTracerProvider, and TestSpan
independently in multiple places. OpenTelemetry already provides testing utils
in their base package. I replaced our Test* objects with an InMemorySpanExporter
which allows us to grab all exported spans and inspect them.
1. In-passing, I realized that for a bit there could be a drift between client
packages that use the new naming conventions and core packages that use the old
naming conventions to pull az.namespace into a span. While it doesn't feel
appropriate to add compatibility code in our core packages, this package will be
in beta for quite some time so I think it's fine to add it here temporarily.