[codespaces] - use prebuilt docker image (#20629)
### Packages impacted by this PR
N/A

### Issues associated with this PR
N/A

### Describe the problem that is addressed by this PR
Our codespaces setup uses a custom Dockerfile in order to configure docker-in-codespaces and install rush.
But nowadays the default image installs docker by default and we can move installing rush to a post-create-script
instead of building the container.

All-in-all this reduces the time to start by over 50%, avoiding the need to build a custom image every time and reusing the
prebuilt default image. It also reduces the amount of code we have to maintain. 

From 1 minute and 7 seconds to 23 seconds before the file tree shows up. I just picked that event arbitrarily.

Resolves #17011