Workaround for LRO polling in phone-numbers sdk
The phoneNumbers API uses LROs with status monitors, as described in the following URL https://github.com/microsoft/api-guidelines/blob/vNext/azure/Guidelines.md#long-running-operations-with-status-monitor.

However, the current implementation of the @azure/core-lro library only supports this pattern if the legacy `azure-asyncoperation` header is present in the response of the operation. Thus, as a workaround to this limitation, a pipeline policy was implemented that adds this header to the responses of LROs, setting it with the same value as the `operation-location` header.