Stop verifying AbortError error message
It turns out that we now have a different error message after switching to Fetch
http client for browser. Previously we were catching the one from xhrHttpClient
but now we get a DOMException from browser.  As error message isn't part of
public Api surface, the verification of error message is removed.