[core-lro] merge azureAsync and location strategies (#20656)
### Packages impacted by this PR
@azure/core-lro 

### Issues associated with this PR
Fixes https://github.com/Azure/azure-sdk-for-js/issues/20647

### Describe the problem that is addressed by this PR
The [vNext draft of the REST guidelines](https://github.com/microsoft/api-guidelines/blob/vNext/azure/Guidelines.md#long-running-operations-with-status-monitor) retires the use of `Azure-AsyncOperation` header in favor of `Operation-Location`. For context, `Azure-AsyncOperation` is mainly used for scenarios where there are two URLs, one for polling (returned in the `Azure-AsyncOperation` header) and another to retrieve the resource being created (returned in the `Location` header).

### What are the possible designs available to address the problem? If there are more than one possible design, why was the one in this PR chosen?
This PR enables clients to use `Operation-Location` for scenarios where `Azure-AsyncOperation` was used before by merging the logic for handling both into one. The merging was done by augmenting the existing logic for handling `Azure-AsyncOperation` as follows: To check if the polling was done (in `isPollingDone`), also check if the response status code was 202 or an unexpected one:
```ts
if (isUnexpectedPollingResponse(rawResponse) || rawResponse.statusCode === 202) {
    return false;
  }
 ```
The new merged logic is moved to `locationPolling.ts` and the old separate logic for handling `Operation-Location` was deleted.

I am open for alternative approaches, but I believe this is the simplest thing we can do.

### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release request PRs)_
N/A

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so, create an Issue in the [Autorest/typescript](https://github.com/Azure/autorest.typescript) repository and link it here)_
- [x] Added a changelog (if necessary)