[EventGrid] Update for March Release (#20693)
This change updates our Swagger API reference and re-runs the code
generation. There are no new events this sprint, however a new value
was added to and enum used by some `Microsoft.Media` events.

As part of the update, we pulled in changes around a new
`aeg-channel-name` header which was added to support a new Event Grid
feature. It is currently unexposed, but we expect to add support in
the April release.