[ServiceBus] Migrate REST-based client to core-rest-pipeline (#19587)
* Move to depend on core v2 libraries.

* Previously WebResourceLike.body is of type any so we were able to assign an object to it, then relies on the internal logic in
`executeAtomXmlOperation()` to stringify it to XML before sending the request. Now that `PipelineRequest.body` is strongly typed, `executeAtomXmlOperation()` is
updated to optionally take the strongly typed request objects for PUT
operations.

* Add a compatibility layer to transform PipelineResponse to HttpResponse like by copying most relevant properties of `WebResourceLike` and add copy of `HttpResponse`

* Bump version

* Stop verifying AbortError error message. It turns out that we now have a different error message after switching to Fetch
http client for browser. Previously we were catching the one from xhrHttpClient
but now we get a DOMException from browser.  As error message isn't part of
public Api surface, the verification of error message is removed.