[ServiceBus] correct types for `propertiesToModify` options (#20577)
We claim that we can take `any` type of property values for properties to be
modified but it's `applicationProperties` part of a Service Bus message that is
modified. That property has a type of `number | boolean | string | Date | null`
and is what the service supports.

This PR corrects the types for these options to align with
`applicationProperties`. Even though it shows as a breaking change, other
unsupported types (e.g., an object literal) would cause service errors so they
never worked.