[Schema Registry Avro] Add an error scenario (#20902)
### Packages impacted by this PR
@azure/schema-registry-avro

### Issues associated with this PR
Fixes https://github.com/Azure/azure-sdk-for-js/issues/20839

### Describe the problem that is addressed by this PR
One of the tests is currently failing because the CI resource only allows up to 25 unique schemas per a schema group but that number is exhausted by another test that checks whether the growth of the serializer's internal cache is bounded. This PR reduces the number of schemas registered in the latter so there is enough room for subsequent tests to register their own schemas.

Furthermore, an additional test is added that checks an error case when deserializing with incompatible schema.

### What are the possible designs available to address the problem? If there are more than one possible design, why was the one in this PR chosen?
N/A

### Are there test cases added in this PR? _(If not, why?)_
Yes. Also see a successful nightly run in https://dev.azure.com/azure-sdk/internal/_build/results?buildId=1439397&view=results

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so, create an Issue in the [Autorest/typescript](https://github.com/Azure/autorest.typescript) repository and link it here)_
- [ ] Added a changelog (if necessary)