[KeyVault] - Add release policy grammar links (#20941)
### Packages impacted by this PR
@azure/keyvault-keys

### Issues associated with this PR
#18400

### Describe the problem that is addressed by this PR
The service team would like to document the release policy grammar, and as such
has asked us to link to their documentation. While the documentation is not live
yet, I did get the content folks to point it to a valid url so that we can
include it in our doc-comments now.