[core-rest-pipeline] Remove authorization header on redirected requests (#21026)
When following a redirect, our redirection policy sent the Authorization header from the original request again in the redirected request. This is a potential security risk, and can also cause issues when the resource at the end of the redirect does not support the authorization method initially provided (as I encountered with ACR in #20957).

This PR causes the redirection policy to remove the header before the request is redirected. This approach is in-line with .NET and Python, which were both already removing this header.