[Tests] upgrade identity dev deps to fix build failures
Identity released version v2.0.4 is not pulling down core-rest-pipeline v1.8.0
because the latter is not released yet so identity credentials find the new
request body type unexpected.  This RP updates the identity dev dependency
version to v2.1.0-beta.2 which is in the repo and references the new
core-rest-pipeline version in the repo.

while fixing container registry I also move the its perf tests to depend on
latest version instead of a pinned beta version at the time when the tests were
created.