Add core-http-compat in Search Documents (#20940)
Search Documents has already migrated to the new Core client and Core rest pipeline libraries, however, this introduced too many breaking changes. In order to minimize the braking changes, core-http-compat was created. This commit regenerates the client with the latest version of autorest, with the core-http-compat-mode flag set to true, and no longer depending on core-http.