[samples] - Fix web-workers sample (#21163)
### Packages impacted by this PR
N/A

### Issues associated with this PR
Fixes #20934

### Describe the problem that is addressed by this PR
A customer reported a few issues:
- At some point when updating our arm-template path we lost a character for
url-encoded / - instead of %2F we only had %2... combined with the followed
character "a" this would end up as the * character instead of the intended
".../a..." character.
- It seems that a recent change in browsers to remove SharedArrayBuffer (see
https://developer.chrome.com/blog/enabling-shared-array-buffer/) has broken
JSDOM's dependencies and by extension JSDOM. This has been fixed upstream. 
For good measure, I just bumped both dependencies.