[KeyVault] - Fix user-agent tests in KeyVault Admin (#21189)
### Packages impacted by this PR
@azure/keyvault-admin 

### Issues associated with this PR
Resolves #21188

### Describe the problem that is addressed by this PR
`KeyVaultClientContext` is a generated file that in core v1 included a
non-exported constant `packageVersion`. Because we have a test that depends on
that constant we always have to update the file to re-export it (as the
generated code does not export it).

Even worse, in coreV2 that constant is not even generated anymore, so the test
is outdated.

This PR fixes all of these issues by examining what we _actually_ care about,
which is that our user-agent includes the correct package version.

This way anyone can regenerate the code at any time without getting weird build
errors.