[eventhub] - Move tracing spec to internal (#21258)
### Packages impacted by this PR
event-hubs

### Issues associated with this PR
Fixes #21226

### Describe the problem that is addressed by this PR
The tracing client test adds an internal reference; however, the test is still
not excluded during min/max due to the nesting of the tracing reference. That
will be fixed separately, but it also makes sense to make this an internal test
as it tests the internals of how we trace.