[event-grid] - Upgrade to core-tracing 1.0 (#21287)
### Packages impacted by this PR
event-grid

### Issues associated with this PR
Resolves #21249

### Describe the problem that is addressed by this PR
As part of a larger effort, this PR upgrades @azure/event-grid to core-tracing
1.0 now that it has GA'd. 

### What are the possible designs available to address the problem? If there are more than one possible design, why was the one in this PR chosen?
One open question was how should I test the
`cloudEventDistributedTracingEnricherPolicy` in this new world. We've done a lot
of work to abstract away from OTel and our test instrumenter no longer creates
headers to avoid polluting the recordings.

But since the policy depends on the traceparent headers to exist we needed to
execute that logic somehow. The solution I landed on is a custom pipeline policy
that puts the two headers this code depends on to exercise that code without
concerning itself with _how_ the headers get there.