[EventGrid] Support Channel Names for Partner Topics (#21411)
The EventGrid service team is adding a new feature called channels
which can be used by Partner Topics (note that "partners" here does
not mean internal Microsoft folks, but rather larger external
customers). The channels act as a sort of namespacing mechanism
(similar to the Domain Topics for topics using the classic
"EventGrid" schema) but are only supported when the schema of the
topic is set to Cloud Events.

Unlike domain topics, where the event payload includes information on
which topic inside the domain to send to, the channel which is
published to is specificed in an `aeg-channel-name` header as part of
the request to send events.

To support this - we add a new property to our options bag,
`channelName` which can be used to specify the channel to use when
sending a set of events. Because this feature is specific to Partner
Topics, we did not try to model it as another named parameter of the
`send` function or create a new method which would be specific to
channels. Most customers won't be using Partner Topics (although,
again, this is a feature that can and is used by external customers)
so we did not want to introduce a new cognitive burden.

We have authored our typescript typings for this function such that
the `channelName` property on the options bag only appears for clients
which are using the CloudEvent schema.

The service team has this feature in Preview today, so we'd like to
release a preview version of the library with support for this feature
that their customers can use.  We've manually validated this against a
partner topic provided by the service team and are working on getting
automated tests that we can record.