[Recorder] Add support for session-level sanitizers (#21533)
The test proxy has support for adding sanitizers at two levels: at the per-recording level and at the session level. Before this PR, the JS recorder client only allowed for sanitizers to be added at the per-recording level (using the `addSanitizers` instance method on the Recorder class). This PR adds a static `Recorder.addSessionSanitizers(...)` method which allows for sanitizers to be set at the session level. These sanitizers will be applied to every recording following the call.