[Text Analytics] Scaling @azure/ai-text-analytics, Part I (#21417)
# Scaling @azure/ai-text-analytics, Part I
This PR reimplements a subset of the public API of the Text Analytics library in a way that avoids the method explosion and the discoverability problems described in https://gist.github.com/deyaaeldeen/40badd8f0460a9a3cd5965b1e4d14adf

## Design Highlights

### One `analyze` to rule them all

https://github.com/Azure/azure-sdk-for-js/blob/23e0919ae7a557a112715a2dbceb0c58a62fde2e/sdk/textanalytics/ai-text-analytics/samples/v5/javascript/analyzeSentimentWithOpinionMining.js#L56-L58

becomes

https://github.com/Azure/azure-sdk-for-js/blob/23e0919ae7a557a112715a2dbceb0c58a62fde2e/sdk/textanalytics/ai-text-analytics/samples/v6-beta/javascript/opinionMining.js#L54-L58

Notice that the type of the action is no longer a standalone method but rather an input to the `analyze` method.

Demo:

https://user-images.githubusercontent.com/6074665/163894076-745be9cb-651a-42da-94d8-c299a44fa0d1.mp4

### Model version and top-level statistics in the response are no longer exposed but still accessible

Model version and top-level statistics are part of the response object and v5 exposes them in the response arrays by attaching them to the array object itself. This works fine but it caused our [documentation to look like that `Array` is part of the client types](https://docs.microsoft.com/en-us/javascript/api/@azure/ai-text-analytics/detectlanguageresultarray?view=azure-node-latest):
![msedge_LFdL6MiUUs](https://user-images.githubusercontent.com/6074665/163894617-4db85932-8954-4690-875a-9d72535e3a1d.png)

They're removed in this design but they can still be accessed as follows:
https://github.com/Azure/azure-sdk-for-js/blob/23e0919ae7a557a112715a2dbceb0c58a62fde2e/sdk/textanalytics/ai-text-analytics/samples/v6-beta/javascript/stats.js#L29-L45

### A batching analyze

Coming soon in a PR near you!