[app-config] Remove custom test rollup config
We are using `nock` in browser test thus need to inject `nock` when rolling up
browser tests. This prevents us from using the centralized rollup config.

For v1 recording, we use `nock` in NodeJS and `nise` in browser. It doesn't make
sense to use `nock` in browser just to run a recorded test.  This PR moves the
test under `node` sub directory and removes the custom test rollup config.