[EventGrid] Fix recorded tests (#21813)
I recently improved the error messages coming out of the recorder in #21575, making the recorder client throw a special error when the recorder failed to find a matching recording for a request. Before that change, the recorder client would yield a standard RestError with a 404 response when a request couldn't be matched to a recording.

Event Grid has a few tests which expect a 404 response from the service. These tests catch and test for a RestError with 404 status to ensure that this response is received. However, during playback mode, the recorder was not finding a recording to match the request with, due to an issue with sanitization. Before my change to the error message, the tests were passing fine (incorrectly), since they were handling the 404 that was coming from the test proxy and not the recording itself. They're now failing since the error that's coming back is different.

The reason that the test proxy could not match the requests to a recording is because of some issues with sanitization of the recording. This change fixes the sanitization and gets the tests passing again, with the RestError now coming from the recording itself giving a 404 response, as was originally intended.