[core-amqp] handle possible `undefined` receiver in timeout handler (#21866)
The live test logs shows that under some circumstance `this.receiver` could be
`undefined`.  We do have code to clean up the timer already so that might have
happened before the clean up and after receiver is closed.

```
     Uncaught TypeError: Cannot read properties of undefined (reading 'address')
      at Receiver.get address [as address] (/Users/runner/work/1/s/common/temp/node_modules/.pnpm/rhea-promise@2.1.0/node_modules/rhea-promise/lib/link.ts:96:24)
      at Timeout._onTimeout (/Users/runner/work/1/s/sdk/core/core-amqp/src/requestResponseLink.ts:158:39)
      at listOnTimeout (node:internal/timers:559:17)
      at processTimers (node:internal/timers:502:7)
```

This PR ensures that we don't throw when `this.receiver` is undefined.  The
address is only used in logging.