[ServiceBus] enable beta-trimmed api-extractor dts rollup
in public-trimmed dts rollup, @beta APIs are removed. Because we only ship
public-trimmed version it prevents us from gathering feedback for @beta APIs.
This PR enable generation of beta-trimmed dts rollup.

We still need a piece in the publishing process to use the proper version
depending on whether we are releasing a beta/dev version or not.