[ServiceBus] don't yield empty page when listing rules in RuleManager
We should stop when there's no more returned rules. I did this previously
probably to support continuationToken as we have to make the last request to
known that there are no more results, and we then set continuationToken to
undefined on this empty page.  Now that the continuationToken is removed, we can
stop yielding empty page.