[KeyVault] keyvault-admin - Don't hard code request scope (#22109)
Currently keyvault-admin falls back to a hard-coded scope when making subsequent requests after the challenge-based authentication flow has been completed. This causes issues when the scope conflicts with the scope returned by the WWW-Authenticate header in the challenge. Hard-coding the scope means that the SDK can't be used in other clouds (e.g., as pointed out by the service team, Dogfood having a scope of https://managedhsm-int.azure-int.net).

This PR updates the logic so that the scope from the original WWW-Authenticate response is used on subsequent requests instead of the hard-coded value.