[core-lro] Update tests to be less verbose (#22377)
From the tests point of view, the path of the initial request is inconsequential and could be anything. I updated the path property to be optional in `LroResponseSpec` and the implementation defaults it to "path". This simplifies the tests a bit because `path` now doesn't have to be specified in the first item in the routes list.