[Language Text] Introduce the library (#22640)
The [archboard recommended](https://github.com/Azure/azure-sdk/issues/4339) creating a new library for the Language Text Analysis service and this PR copies the codebase of @azure/ai-text-analytics into the new @azure/ai-language-text. The only other change the PR makes is updating all the names/references accordingly. Please note that link verification will fail for many links that don't exist on the main branch yet and I plan to override it to get the PR merged.