[ServiceBus] Fix bug in converting ServiceBusMessage to RheaMessage (#22665)
when the `ServiceBusMessage` has `timeToLive` property set.

The problem is that we are adding a number to a `Date` object, resulting in an
invalid date value for `absolute_expiry_time`. This PR corrects it by calling
`getTime()` on the `Date` object first before adding the ttl number.

A unit test is added and the sendMessages sample is updated to show how to set
`timeToLive`.