[core] Address Trusted Types compliance issue
Trusted Types is/are a set of web APIs that give us the ability to control and
restrict data that flows into dangerous DOM APIs in our apps. Main goal of this
mechanism is to prevent successful exploitation of HTML and DOM-based
Cross-Site-Scripting (XSS) vulnerabilities.

  https://github.com/w3c/webappsec-trusted-types

Our code does not parse HTML strings. The parsed DOM object is not exposed to
outside world either. This PR adds trusted types policies to make the parsing
code compliant at running time. Consumers of our packages can review and add the
policy to their allowed list.