[core-amqp][test] fix test issue comparing two timestamps (#22769)
These two tests compare two expiry values from two consecutive Date.now()
calls (the first in the test and the second in the library code) and expect them
to be equal:

    Math.floor(Date.now() / 1000) + 3600

However, it happens that the elapsed time between the two Date.now() call
sometimes can fail the expectation.

> AssertionError: expected 1659139755 to equal 1659139754

This PR change the verification to allow max of 2 milliseconds delta between the
two timestamps.

While at this I also updated the nyc reporters to more useful ones. Previous
ones just print out lines of data that not meant to be human-understandable.