Replace `beginGet*Batch` with `resume*Batch`
We provide a `beginGet*Batch` method to resume the previous
request previously. This is not ideal because it need to get a
`batchId` which is store in the internal state of the LRO. Once
the LRO brings changes, our code will be broken.

We can achieve the same thing by using serialized state.
This change aims to leverage this approach and create convenience
methods `resume*Batch` for users