[engsys] update dep core-util version to `^1.1.1`
to ensure packages can get the `delay()` refactoring.

Without this change our `js - core` nightly is failing. It would fail the
min-max testing as well.

What's happening is that our nightly pipeline does something special: for
example rewrite our @azure/core-util package versions to be like
"1.1.1-alpha.20220909.2" so we can publish nightly alpha package of this version
to npmjs.org.

If our other packages depend on the same minimum version (^1.1.1) then we also
update the version to be ">=1.1.1-alpha <1.1.1-alphb" then rush would be able to
find this version in the repository and we are building and testing against the
version to be published.

However, if the version doesn't match (^1.1.0 in this case), rush would not
modify the dep version, and would get v1.1.0 from npmjs instead.

Using "^1.1.1" is also more correct for end users, as v1.1.0 satisfies the sem
version ^1.1.0 but doesn't work.