[Language Text] Correct implementation of auto language detection (#23683)
The old implementation was an added type constraint but that was not correct because the types used for instantiation didn't respect the constraint. This PR replaces the constraint with an intersection type instead.

This implementation makes sure that `detectedLanguage` is part of the output models for `beginAnalyzeBatch` but not for the corresponding ones for `analyze`.

The PR also adds `isLanguageDefaulted` property which is returned by the service but not described in the swagger.